server/requirements.txt
APScheduler==3.10.4
asyncio==3.4.3
Authlib==1.3.0
bcrypt==4.1.3
DateTime==5.5
email_validator==2.1.1
Jinja2==3.1.3
fastapi==0.110.3
loguru==0.7.2
pillow==10.3.0
pydantic-settings==2.2.1
pymongo==4.7.1
python-multipart==0.0.9
requests==2.31.0
# tqdm==4.66.4
uuid==1.30
uvicorn==0.28.1